# Avoid copying nodes where possible