--- a/internal/scheduler/preempting_queue_scheduler_test.go
+++ b/internal/scheduler/preempting_queue_scheduler_test.go
@@ -1254,7 +1254,7 @@ func TestPreemptingQueueScheduler(t *testing.T) {
 	}
 	for name, tc := range tests {
 		t.Run(name, func(t *testing.T) {
-			nodeDb, err := CreateNodeDb(tc.Nodes)
+			nodeDb, err := NewNodeDb(tc.Nodes)
 			require.NoError(t, err)
 
 			// Repo. for storing jobs to be queued.
@@ -1577,7 +1577,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 				priorityFactorByQueue[queue] = float64(rand.Intn(tc.MaxPriorityFactor-tc.MinPriorityFactor+1) + tc.MinPriorityFactor)
 			}
 
-			nodeDb, err := CreateNodeDb(tc.Nodes)
+			nodeDb, err := NewNodeDb(tc.Nodes)
 			require.NoError(b, err)
 			jobRepo := NewInMemoryJobRepository(testfixtures.TestPriorityClasses)
 
@@ -1643,7 +1643,7 @@ func BenchmarkPreemptingQueueScheduler(b *testing.B) {
 				require.NoError(b, err)
 				nodesById[nodeId] = node
 			}
-			nodeDb, err = CreateNodeDb(maps.Values(nodesById))
+			nodeDb, err = NewNodeDb(maps.Values(nodesById))
 			require.NoError(b, err)
 
 			allocatedByQueueAndPriorityClass := sctx.AllocatedByQueueAndPriority()