Concretely, there are two situations in which we can avoid intermediate copies of nodes:

1. When constructing a `NodeDb`, we were making a separate call to `BindJobToNode` for each already-running job (which meant that we had to copy a `Node` for each already-running job); instead, we now first group the slice of already-running jobs according to the nodes that they're running on, and then only copy each node at most once.

   I added a benchmark for this case, which is much faster now (after b3ac7c46c8e1e1e3e8d7a418cb7990bec96b8e51, it runs in a few seconds on my machine; I don't know how long it took before that because I haven't had the patience to let it run for more than 15 minutes).
2. When evicting jobs at the start of the scheduling round, we were making a separate call to `EvictJobFromNode` for each preemptible job (which, again, meant that we had to copy a `Node` for each preemptible job); instead, we now evict all preemptible jobs on a given node in one go.

   I already attempted to make this change in #2590, but this PR comes with a cleaner API (no in-place node operations are exported anymore). As in the previous PR, this change makes a noticeable difference when running `go test ./internal/scheduler -bench='^BenchmarkPreemptingQueueScheduler$' -count=10 -run='^$'`:

   ```
   goos: darwin
   goarch: amd64
   pkg: github.com/armadaproject/armada/internal/scheduler
   cpu: Intel(R) Core(TM) i5-8210Y CPU @ 1.60GHz
                                                               │   old.txt    │                new.txt                │
                                                               │    sec/op    │    sec/op      vs base                │
   PreemptingQueueScheduler/100_nodes_1_queue_32000_jobs-4       541.0m ±  8%   414.8m ± 137%  -23.33% (p=0.023 n=10)
   PreemptingQueueScheduler/100_nodes_10_queues_32000_jobs-4     802.0m ± 17%   529.3m ±   9%  -34.00% (p=0.000 n=10)
   PreemptingQueueScheduler/1000_nodes_1_queue_320000_jobs-4      7.651 ± 35%    5.123 ±  12%  -33.05% (p=0.001 n=10)
   PreemptingQueueScheduler/1000_nodes_10_queues_320000_jobs-4    5.572 ± 47%    4.132 ±  25%  -25.85% (p=0.001 n=10)
   PreemptingQueueScheduler/1_node_1_queue_320_jobs-4            4.762m ±  6%   3.525m ±   4%  -25.99% (p=0.000 n=10)
   PreemptingQueueScheduler/1_node_10_queues_320_jobs-4          5.792m ±  6%   4.256m ±   5%  -26.51% (p=0.000 n=10)
   PreemptingQueueScheduler/10_nodes_1_queue_3200_jobs-4         49.45m ±  2%   38.04m ±   3%  -23.08% (p=0.000 n=10)
   PreemptingQueueScheduler/10_nodes_10_queues_3200_jobs-4       62.23m ±  6%   46.07m ±   2%  -25.97% (p=0.000 n=10)
   geomean                                                       188.1m         136.7m         -27.32%
   ```

┆Issue is synchronized with this [Jira Task](https://gr-oss.atlassian.net/browse/BATCH-410) by [Unito](https://www.unito.io)